# Improve Regex UpdateBumpalong optimization for non-atomic and lazy loops