Fixes #62676 
Fixes #62677

In addition to adding more tests for lazy loops, this PR makes two improvements for "update bumpalong".  For the interpreter/compiler/source generator, the scan loop finds the next likely position for a match, then runs the engine at that position, then finds the next likely position for a match, then runs the engine at that position, etc.  Normally after a failed match the scan loop just "bumps" the position by 1 beyond where it previously started.  But it's possible for the matching logic to update that bumpalong position further when it sees that starting from an earlier position would be pointless.  In .NET 5, we added such an optimization for the case where an expression begins with a greedy loop with an infinite upper bound, primarily focused on atomic loops (but we'd add the UpdateBumpalong node for non-atomic as well).  In such cases, we don't need to repeat the matching at any position within the space explored by that loop, since we'll just be re-evaluating the same conditions, and so we bump the runtextpos position to whatever it is after the loop.

That works well for atomic loops, where we won't backtrack.  But for non-atomic loops, this optimization ended up being a nop: in the case of a failed match, we'd backtrack into the loop and back off to a lower position, and then dutifully store that position... that means the runtextpos would initially be bumped to the appropriate place, but by the end of the matching would have worked its way back most or all the way to where we started.  The simple fix here isn't to just always set the position to the one after the loop but instead only set it if it's greater than what's currently stored.

We can also insert UpdateBumpalong for non-atomic lazy loops, since in the case of failure (which is the only case where UpdateBumpalong has an impact), a greedy loop and a lazy loop end up exploring the same search space, just in a different direction.

Example artificial microbenchmarks to highlight the possible impact:
```C#
using System.Linq;
using System.Text.RegularExpressions;
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;

[MemoryDiagnoser]
public class Program
{
    public static void Main(string[] args) => BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run(args);

    private Regex _greedy = new Regex(".*abcd", RegexOptions.Compiled);
    private Regex _lazy = new Regex(".*?abcd", RegexOptions.Compiled);
    private string _input = string.Concat(Enumerable.Repeat("abcDefghijklmnopqrstuvwxyz", 1000));

    [Benchmark] public void Greedy() => _greedy.IsMatch(_input);
    [Benchmark] public void Lazy() => _lazy.IsMatch(_input);
}
```

| Method |         Toolchain |            Mean | Ratio |
|------- |------------------ |----------------:|------:|
| Greedy | \main\corerun.exe | 1,287,192.23 us | 1.000 |
| Greedy |   \pr\corerun.exe |        85.61 us | 0.000 |
|        |                   |                 |       |
|   Lazy | \main\corerun.exe |   555,555.78 us | 1.000 |
|   Lazy |   \pr\corerun.exe |        42.33 us | 0.000 |